Update README.md
Added an additional troubleshooting section entitled "Changes to framework aren't being shown in apps".  This was a problem that drove me crazy for hours.  I'm still not sure of the actual cause.  There didn't seem to be any trace of the old version of the framework, but the app still used it on simulator and device.